Add shift out add from master list button
