+++ packages/requisitions/src/RequestRequisition/DetailView/AppBarButtons/AppBarButtons.tsx
@@ -6,19 +6,18 @@ import {
   Grid,
   useDetailPanel,
   useTranslation,
-  useToggle,
   PrinterIcon,
   LoadingButton,
   ReportCategory,
 } from '@openmsupply-client/common';
 import {
-  MasterListSearchModal,
   ReportRowFragment,
   ReportSelector,
   usePrintReport,
 } from '@openmsupply-client/system';
-import { useAddFromMasterList, useRequest } from '../../api';
+import { useIsRequestDisabled, useRequest } from '../../api';
 import { UseSuggestedQuantityButton } from './UseSuggestedQuantityButton';
+import { AddFromMasterListButton } from './AddFromMasterListButton';
 
 interface AppBarButtonProps {
   isDisabled: boolean;
@@ -26,13 +25,11 @@ interface AppBarButtonProps {
 }
 
 export const AppBarButtonsComponent: FC<AppBarButtonProps> = ({
-  isDisabled,
   onAddItem,
 }) => {
-  const { addFromMasterList } = useAddFromMasterList();
+  const isDisabled = useIsRequestDisabled();
   const { OpenButton } = useDetailPanel();
   const t = useTranslation('distribution');
-  const modalController = useToggle();
   const { data } = useRequest();
   const { print, isPrinting } = usePrintReport();
 
@@ -45,20 +42,8 @@ export const AppBarButtonsComponent: FC<AppBarButtonProps> = ({
     <AppBarButtonsPortal>
       <Grid container gap={1}>
         <UseSuggestedQuantityButton />
-        <ButtonWithIcon
-          disabled={isDisabled}
-          Icon={<PlusCircleIcon />}
-          label={t('button.add-from-master-list')}
-          onClick={modalController.toggleOn}
-        />
-        <MasterListSearchModal
-          open={modalController.isOn}
-          onClose={modalController.toggleOff}
-          onChange={masterList => {
-            modalController.toggleOff();
-            addFromMasterList(masterList);
-          }}
-        />
+        <AddFromMasterListButton />
+
         <ButtonWithIcon
           disabled={isDisabled}
           label={t('button.add-item')}
